+++ compiler-plugin/src/main/java/io/ballerina/stdlib/sql/compiler/SQLDiagnosticsCodes.java
@@ -20,7 +20,6 @@ package io.ballerina.stdlib.sql.compiler;
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 
 import static io.ballerina.tools.diagnostics.DiagnosticSeverity.ERROR;
-import static io.ballerina.tools.diagnostics.DiagnosticSeverity.HINT;
 
 /**
  * Enum class to hold SQL module diagnostic codes.
@@ -28,7 +27,7 @@ import static io.ballerina.tools.diagnostics.DiagnosticSeverity.HINT;
 public enum SQLDiagnosticsCodes {
     SQL_101("SQL_101", "invalid value: expected value is greater than one", ERROR),
     SQL_102("SQL_102", "invalid value: expected value is greater than zero", ERROR),
-    SQL_103("SQL_103", "invalid value: expected value is greater than or equal to 30", ERROR),
+    SQL_103("SQL_103", "invalid value: expected value is either 0 or greater than or equal to 30", ERROR),
 
     // Out parameter return type validations diagnostics
     SQL_201("SQL_201", "invalid value: expected value is array", ERROR),
@@ -43,9 +42,7 @@ public enum SQLDiagnosticsCodes {
     SQL_221("SQL_221", "invalid value: expected value is any one of boolean, int or string", ERROR),
     SQL_222("SQL_222", "invalid value: expected value is any one of time:Date, int or string", ERROR),
     SQL_223("SQL_223", "invalid value: expected value is any one of time:TimeOfDay, int or string", ERROR),
-    SQL_231("SQL_231", "invalid value: expected value is any one of time:Civil, time:Utc, int or string", ERROR),
-
-    SQL_901("SQL_901", "parameter 'typeDesc' should be explicitly passed when the return data is ignored", HINT);
+    SQL_231("SQL_231", "invalid value: expected value is any one of time:Civil, time:Utc, int or string", ERROR);
 
     private final String code;
     private final String message;