+++ changelog.md
@@ -9,6 +9,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- [Support retrieval of enum types](https://github.com/ballerina-platform/ballerina-standard-library/issues/4588)
+
+## [1.9.0] - 2023-06-01
+
+### Changed
+- [Improve documentation regard `sql:Column` annotation](https://github.com/ballerina-platform/ballerina-standard-library/issues/4134)
+- [Handle null error messages from underlying drivers](https://github.com/ballerina-platform/ballerina-standard-library/issues/4200)
+- [Make `sql:Client` isolated](https://github.com/ballerina-platform/ballerina-standard-library/issues/4455)
+
+## [1.7.1] - 2023-03-09
+
+### Changed
+
+- [Optimise batchExecute with a batch size of 1000](https://github.com/ballerina-platform/ballerina-standard-library/issues/4129)
+
+## [1.7.0] - 2023-02-20
+
+### Changed
+
+- [Remove SQL_901 diagnostic hint](https://github.com/ballerina-platform/ballerina-standard-library/issues/3609)
+- [Enable non-Hikari logs](https://github.com/ballerina-platform/ballerina-standard-library/issues/3763)
+
+## [1.6.2] - 2023-02-09
+
+### Changed
+
+- [Improve API docs based on Best practices](https://github.com/ballerina-platform/ballerina-standard-library/issues/3857)
+- [Fix SQL compiler plugins failure when the diagnostic code is null](https://github.com/ballerina-platform/ballerina-standard-library/issues/4054)
+
+## [1.6.1] - 2022-12-22
+
+### Changed
+- [Enable non-Hikari logs](https://github.com/ballerina-platform/ballerina-standard-library/issues/3763)
+
+## [1.6.0] - 2022-11-29
+
+### Changed
+- [Updated API Docs](https://github.com/ballerina-platform/ballerina-standard-library/issues/3463)
+- [Fix unable to set unlimited lifetime (0) to ballerina.sql.maxConnectionLifeTime](https://github.com/ballerina-platform/ballerina-standard-library/issues/3657)
+- [Improve error message on client connection failure](https://github.com/ballerina-platform/ballerina-standard-library/issues/3648)
 
 ## [1.5.0] - 2022-09-08